monolith_app/app/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
sqlalchemy==1.4.39
psycopg2-binary==2.9.3
kafka-python==2.0.2
python-dotenv==1.0.0
tenacity==9.0.0

aiokafka==0.8.0
asyncpg==0.24.0
sqlalchemy[asyncio]==1.4.54

